+++ MCFCRT/src/stdc/string/strchr.c
@@ -18,7 +18,7 @@ char *strchr(const char *s, int c){
 	__MCFCRT_xmmsetz(xz);
 	__m128i xc[1];
 	__MCFCRT_xmmsetb(xc, (uint8_t)c);
-	ptrdiff_t shift = (const char *)s - rp;
+	int shift = (int)((const char *)s - rp);
 	uint32_t skip = (uint32_t)-1 << shift;
 	for(;;){
 		__m128i xw[2];
